requirements.txt
cryptography==3.4.7
django==3.2.6
social-auth-app-django==4.0.0
python-jose==3.2.0
mysqlclient==2.0.3
djangorestframework==3.12.4
prometheus-client==0.10.1
django-prometheus==2.1.0
gunicorn==20.1.0
django-cors-headers==3.13.0
